requirements.txt
Django>=5.0,<6
djangorestframework>=3.15
celery[redis]>=5.4
django-celery-beat>=2.7
django-celery-results>=2.5
redis>=5
psycopg[binary]>=3.2
requests>=2.31
drf-spectacular>=0.27
pycountry>=23.12.11
python-dotenv>=1.0
pandas>=2.0
